로그인, 로그아웃, 작성자, 권한체크 구현

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -1,7 +1,5 @@
 package com.example.demo.controller;
 
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -10,7 +8,11 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import com.example.demo.service.ArticleService;
 import com.example.demo.util.Ut;
 import com.example.demo.vo.Article;
+import com.example.demo.vo.Member;
 import com.example.demo.vo.ResultData;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 
 @Controller
 public class UsrArticleController {
@@ -22,18 +24,31 @@ public class UsrArticleController {
 	// 액션 메서드 (컨트롤러 메서드)
 	@RequestMapping("/usr/article/doWrite")
 	@ResponseBody
-	public ResultData doAdd(String title, String body) {
+	public ResultData doWrite(HttpServletRequest request, String title, String body) {
+		HttpSession session = request.getSession();
+		if (Ut.isEmptyOrNull(title)) {
+			return ResultData.from("F-1", "제목을 입력하세요");
+		}
 
-		int id = articleService.writeArticle(body, title);
+		if (Ut.isEmptyOrNull(body)) {
+			return ResultData.from("F-2", "내용을 입력하세요");
+		}
+
+		if (session.getAttribute("loginedMember") == null) {
+			return ResultData.from("F-3", "로그인 후 이용해주세요");
+		}
+
+		int memberId = (int) session.getAttribute("loginedMemberId");
+		int id = articleService.writeArticle(body, title, memberId);
 		Article article = articleService.getArticleById(id);
-		return ResultData.from("S-4", Ut.f("%d번 게시글입니다", id), article);
+		return ResultData.from("S-1", Ut.f("%d번 게시글입니다", id), article);
 	}
 
 	@RequestMapping("/usr/article/getArticles")
 	@ResponseBody
 	public ResultData getArticles() {
 
-		return ResultData.from("S-5", Ut.f("게시글입니다"), articleService.getArticles());
+		return ResultData.from("S-1", Ut.f("게시글입니다"), articleService.getArticles());
 	}
 
 	@RequestMapping("/usr/article/getArticle")
@@ -52,31 +67,46 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData doDelete(HttpServletRequest request, int id) {
+		HttpSession session = request.getSession();
+		if (session.getAttribute("loginedMember") == null) {
+			return ResultData.from("F-2", "로그인 후 이용해주세요");
+		}
 		Article article = articleService.getArticleById(id);
 
 		if (article == null) {
-			return ResultData.from("F-2", Ut.f("%d번 게시글은 없습니다", id));
+			return ResultData.from("F-1", Ut.f("%d번 게시글은 없습니다", id));
 		}
 
+		if (article.getMemberId() != (int) session.getAttribute("loginedMemberId")) {
+			return ResultData.from("F-3", Ut.f("권한이 없습니다."));
+		}
 		articleService.deleteArticle(id);
 
-		return ResultData.from("S-2", Ut.f("%d번 게시글이 삭제되었습니다.", id), article);
+		return ResultData.from("S-1", Ut.f("%d번 게시글이 삭제되었습니다.", id), article);
 
 	}
 
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(int id, String title, String body) {
+	public ResultData doModify(HttpServletRequest request, int id, String title, String body) {
+		HttpSession session = request.getSession();
+		if (session.getAttribute("loginedMember") == null) {
+			return ResultData.from("F-2", "로그인 후 이용해주세요");
+		}
 		Article article = articleService.getArticleById(id);
 
 		if (article == null) {
-			return ResultData.from("F-3", Ut.f("%d번 게시글은 없습니다", id));
+			return ResultData.from("F-1", Ut.f("%d번 게시글은 없습니다", id));
+		}
+
+		if (article.getMemberId() != (int) session.getAttribute("loginedMemberId")) {
+			return ResultData.from("F-3", Ut.f("권한이 없습니다."));
 		}
 
 		articleService.modifyArticle(id, title, body);
 
-		return ResultData.from("S-3", Ut.f("%d번 게시글이 수정되었습니다.", id), article);
+		return ResultData.from("S-1", Ut.f("%d번 게시글이 수정되었습니다.", id), article);
 
 	}
